사용자 인증 구현 성공

+++ src/service/user.service.ts
@@ -9,7 +9,7 @@ export class UserService {
     constructor(@InjectRepository(User) private userRepository: Repository<User>) {
         this.userRepository = userRepository;
     }
-      
+
     //User 리스트 조회  
     findAll(): Promise<User[]> {
         return this.userRepository.find();
@@ -19,13 +19,22 @@ export class UserService {
         return this.userRepository.findOne({ user_id: id });
     }
 
-    // 유저 저장
-    async saveUser(user: User): Promise<void> {
-        await this.userRepository.save(user);
+    // 유저 회원가입
+    async createUser(user : User): Promise<User> {
+        const isExist = await this.findOne(user.user_id);
+        if(isExist) {
+            return null;
+        }        
+        return await this.userRepository.save(user);
     }
 
     // 유저 삭제
-    async deleteUser(id: string): Promise<void> {
-        await this.userRepository.delete({ user_id: id });
+    async deleteUser(id: string): Promise<boolean> {
+        const isExist = await this.findOne(id);
+        if(isExist) {
+            await this.userRepository.delete({ user_id: id });
+            return true;
+        }  
+        return false;
     }
 }
